+++ lib/kumade.rb
@@ -1,6 +1,7 @@
 require 'rake'
 require 'thor'
 
+require 'kumade/base'
 require 'kumade/git'
 require 'kumade/deployer'
 require 'kumade/runner'
